// src/app/admin/tags/tag.type.ts
export type AdminTagItem = {
  id: string;
  slug: string;
  sortOrder: number;
  isActive: boolean;
  nameVi: string;
  nameEn: string;
  slugVi: string;
  slugEn: string;
  descriptionVi: string;
  descriptionEn: string;
  postsCount: number;
};

export type AdminTagDetail = {
  id: string;
  slug: string;
  sortOrder: number;
  isActive: boolean;
  nameVi: string;
  nameEn: string;
  slugVi: string;
  slugEn: string;
  descriptionVi: string;
  descriptionEn: string;
};